+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed a small bug that lets you open non-markdown files as roots.
 - You can now copy selected Markdown text as HTML! Just press `Cmd/Ctrl+Alt+C` instead of `Cmd/Ctrl+C`.
 - Added an online-check. From now on, if you are offline, Zettlr won't show you ugly error messages, but a simple notification that you are, in fact, offline and Zettlr can't check for updates.
+- Improved footnote placement.
 
 ## Under the hood
 
@@ -23,6 +24,7 @@
 - Changed resizing constraints: Editor can now have 90 percent width at maximum.
 - Fixed a logical error in `ZettlrConfig` that allowed non-markdown-files to be opened as root files.
 - Buffered the update check with an online-check. Renamed the original `check()` to `_fetchReleases()`.
+- Fixes to the footnote placement.
 
 # 0.18.0